Winkel und Geschwindigkeiten sind jetzt float
- convertermethoden geben jetzt float zurück
- das model speichert die Geschwindigkeiten als float
- werte im OffsetConfig auf float geändert
- ARDroneCommander sendet move-Befehle jetzt in float

+++ QuadroMotion/src/com/quadromotion/config/OffsetConfig.java
@@ -2,29 +2,29 @@ package com.quadromotion.config;
 
 public class OffsetConfig {
 	
-	public static final double MAX_ANGLE_X = 80;
-	public static final double MAX_SPEED_X = 50;
-	public static final double SPEED_OFFSET_X = 10;
-	public static final double ANGLE_OFFSET_X = 15;
-	public static final double FUNCTION_EXP_X = 2.1;
+	public static final float MAX_ANGLE_X = 80;
+	public static final float MAX_SPEED_X = 50;
+	public static final float SPEED_OFFSET_X = 10;
+	public static final float ANGLE_OFFSET_X = 15;
+	public static final float FUNCTION_EXP_X = 2.1F;
 	
-	public static final double MAX_ANGLE_Y = 80;
-	public static final double MAX_SPEED_Y = 50;
-	public static final double SPEED_OFFSET_Y = 10;
-	public static final double ANGLE_OFFSET_Y = 15;
-	public static final double FUNCTION_EXP_Y = 2.1;
+	public static final float MAX_ANGLE_Y = 80;
+	public static final float MAX_SPEED_Y = 50;
+	public static final float SPEED_OFFSET_Y = 10;
+	public static final float ANGLE_OFFSET_Y = 15;
+	public static final float FUNCTION_EXP_Y = 2.1F;
 	
-	public static final double MAX_ANGLE_Z = 80;
-	public static final double MAX_SPEED_Z = 50;
-	public static final double SPEED_OFFSET_Z = 10;
-	public static final double ANGLE_OFFSET_Z = 15;
-	public static final double FUNCTION_EXP_Z = 2.1;
+	public static final float MAX_ANGLE_Z = 80;
+	public static final float MAX_SPEED_Z = 50;
+	public static final float SPEED_OFFSET_Z = 10;
+	public static final float ANGLE_OFFSET_Z = 15;
+	public static final float FUNCTION_EXP_Z = 2.1F;
 	
-	public static final double MAX_ANGLE_SPIN = 80;
-	public static final double MAX_SPEED_SPIN = 50;
-	public static final double SPEED_OFFSET_SPIN = 10;
-	public static final double ANGLE_OFFSET_SPIN = 15;
-	public static final double FUNCTION_EXP_SPIN = 2.1;
+	public static final float MAX_ANGLE_SPIN = 80;
+	public static final float MAX_SPEED_SPIN = 50;
+	public static final float SPEED_OFFSET_SPIN = 10;
+	public static final float ANGLE_OFFSET_SPIN = 15;
+	public static final float FUNCTION_EXP_SPIN = 2.1F;
 	
 	/**
 	 *
