+++ QuadroMotion/src/com/quadromotion/controller/IARDroneCommander.java
@@ -14,7 +14,7 @@ public interface IARDroneCommander {
 	 * @param speedZ the speed in direction Z, can be positive or negative
 	 * @param speedSpin the speed to spin, can be positive or negative
 	 */
-	public void moveDrone(int speedX, int speedY, int speedZ, int speedSpin);
+	public void moveDrone(float speedX, float speedY, float speedZ, float speedSpin);
 	
 	/**
 	 * This method is needed to send the hover command
